% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_with_translated_texts.R
\name{complete_with_translated_texts}
\alias{complete_with_translated_texts}
\title{Translate text if necessary and place it in the data tibble}
\usage{
complete_with_translated_texts(
  text_to_complete,
  text_to_translate,
  whether_to_translate,
  long = FALSE
)
}
\arguments{
\item{text_to_complete}{the vector of texts to complete with translated texts (the elements such that whether_to_translate is FALSE are left unchanged)}

\item{text_to_translate}{the text to translate}

\item{whether_to_translate}{whether or not to translate the text to English (can be FALSE if the text is already in English for instance).}

\item{long}{whether the text is long (polyglotr's google_translate_long() function will be used if long is TRUE)}
}
\value{
a tibble
}
\description{
Translate text if necessary and place it in the data tibble
}
\examples{
tib=tibble::tibble(text_original=c("Do not translate","Oui celui-là, oui", "Allons-y"),
translate=c(FALSE,TRUE,TRUE))
library(dplyr)
tib \%>\%
  mutate(text_trans=complete_with_translated_texts(text_original,text_original,translate))
}
