% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Description}
\usage{
cross(collection, data_dir, type = "people")
}
\arguments{
\item{collection}{a collection name}

\item{data_dir}{the path to data directory}

\item{type}{whether to group by "people" or "labs"}

\item{method}{method to shorten names. Defaults to "shortest"}
}
\value{
a tibble which crosses people or labs to tally number of collaborations
}
\description{
Description
}
\examples{
HALtere::cross("BIOEENVIS","data", type="people")
}
