% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_collection.R
\name{extract_collection}
\alias{extract_collection}
\title{Creates file publications.RDS as direct export from HAL API in data_dir}
\usage{
extract_collection(
  collection,
  data_dir = "data",
  nmax = +Inf,
  hal_variable = "collCode_s"
)
}
\arguments{
\item{collection}{name of a HAL collection}

\item{data_dir}{path to data directory. Defaults to "data".}

\item{hal_variable}{the HAL variable used to retrieve items. Can be collCode_s (for collections), authIdHal_s (for person)}

\item{n_max}{the maximum number of items to retrieve (this argument is useful for testing purposes). Defaults to 200.}
}
\value{
a tibble
}
\description{
Creates file publications.RDS as direct export from HAL API in data_dir
}
\examples{
publis=extract_collection("EVS_UMR5600", nmax=30)
}
