% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_publications.R
\name{extract_ith_slice_of_publications}
\alias{extract_ith_slice_of_publications}
\title{Extract the publication slice going from i*100 to i*100+100}
\usage{
extract_ith_slice_of_publications(i, hal_variable, identifier, nmax = +Inf)
}
\arguments{
\item{i}{the slice number}

\item{hal_variable}{the HAL variable used to retrieve items. Can be collCode_s (for collections), authIdHal_s (for person)}

\item{identifier}{identifier of the collection/person/lab}

\item{nmax}{maximum number of publications to extract}
}
\value{
a data.frame containing a slice of publications
}
\description{
Extract the publication slice going from i*100 to i*100+100
}
