% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_publications.R
\name{extract_publications}
\alias{extract_publications}
\title{Creates file publications.RDS as direct export from HAL API in data_dir}
\usage{
extract_publications(
  custom_name,
  hal_variable,
  identifier,
  data_dir = "data_HALtere",
  nmax = +Inf
)
}
\arguments{
\item{custom_name}{name to give to the list of publications. If NA (default) the directory will be named after the used HAL identifier.}

\item{hal_variable}{the HAL variable used to retrieve items. Can be collCode_s (for collections), authIdHal_s (for person)}

\item{identifier}{the identifier (going with hal_variable) used to retrieve HAL items.}

\item{data_dir}{path to data directory. Defaults to "data_HALtere".}

\item{n_max}{the maximum number of items to retrieve (this argument is useful for testing purposes). Defaults to 200.}
}
\value{
a tibble
}
\description{
Creates file publications.RDS as direct export from HAL API in data_dir
}
\examples{
publis=extract_publications("EVS_UMR5600", nmax=30)
}
