% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hal_labels.R
\name{get_hal_labels}
\alias{get_hal_labels}
\title{Retrieve a human-readable HAL label from an identifier}
\usage{
get_hal_labels(identifier, identifier_type)
}
\arguments{
\item{identifier}{identifier value}

\item{identifier_type}{HAL field used as identifier}
}
\value{
character vector
}
\description{
Retrieve a human-readable HAL label from an identifier
}
