% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_all_data.R
\name{prepare_all_data}
\alias{prepare_all_data}
\title{Get all HALtere data}
\usage{
prepare_all_data(custom_name, hal_variable, identifier)
}
\arguments{
\item{custom_name}{name to give to the list of publications. If NA (default) the directory will be named after the used HAL identifier.}

\item{hal_variable}{the HAL variable used to retrieve items. Can be collCode_s (for collections), authIdHal_s (for person)}

\item{identifier}{the identifier (going with hal_variable) used to retrieve HAL items.}
}
\value{
the characteristics of datasets retrieved for this list of publications
}
\description{
Get all HALtere data
}
\examples{
prepare_all_data("lise-vaudor",hal_variable="person")
}
