% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_all_data.R
\name{prepare_HALtere_directory}
\alias{prepare_HALtere_directory}
\title{Prepare line of HALtere directories-related data}
\usage{
prepare_HALtere_directory(identifier, hal_variable, custom_name = NA)
}
\arguments{
\item{identifier}{the identifier (going with hal_variable) used to retrieve HAL items.}

\item{hal_variable}{the HAL variable used to retrieve items. Can be collCode_s (for collections), authIdHal_s (for person)}

\item{custom_name}{name to give to the list of publications. If NA (default) the directory will be named after the used HAL identifier.}
}
\value{
one line of dataset regarding a list of publications
}
\description{
Prepare line of HALtere directories-related data
}
\examples{
prepare_HALtere_directory(custom_name="Lise Vaudor",
                         identifier="lise-vaudor",
                         hal_variable="authIdHal_s")
}
