% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null_with_na.R
\name{replace_null_with_na}
\alias{replace_null_with_na}
\title{Replaces nulls in a list with NA}
\usage{
replace_null_with_na(list_with_nulls)
}
\arguments{
\item{list_with_nulls}{a list comprising nulls}
}
\value{
list comprising NAs
}
\description{
Replaces nulls in a list with NA
}
\examples{
replace_null_with_na(list(0,33,"a", NULL,"b",NA,NULL))
}
