% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_publications.R
\name{retry_skip}
\alias{retry_skip}
\title{Retry to run a code}
\usage{
retry_skip(expr, max_try = 3, wait = 0)
}
\arguments{
\item{max_try}{maximum number of trials. Defaults to 3}

\item{wait}{the number of seconds to wait between trials. Defaults to 0.}
}
\description{
Retry to run a code
}
