% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_authors.R
\name{show_ref_authors}
\alias{show_ref_authors}
\title{Produces a simplified table of ref_authors}
\usage{
show_ref_authors(publications, data_ref_authors)
}
\arguments{
\item{publications}{a tibble produced with the extract_publications() function}

\item{data_ref_authors}{a tibble produced with the tidy_ref_authors() function}
}
\value{
a tibble with simplified ref_authors information
}
\description{
Produces a simplified table of ref_authors
}
\examples{
data=extract_publications("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)
show_ref_authors(publications=data, data_ref_authors=data_ref_authors)
}
