% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_text.R
\name{tidy_text}
\alias{tidy_text}
\title{Produces data_text, which is derived from a publications tibble}
\usage{
tidy_text(custom_name, data_dir = "data", column)
}
\arguments{
\item{custom_name}{a custom_name name}

\item{data_dir}{the path to data directory}
}
\value{
a tibble
}
\description{
Produces data_text, which is derived from a publications tibble
}
\examples{
data_text=tidy_text("BIOEENVIS", data=data_dir,column="text")
}
