% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_authors.R
\name{tidy_words}
\alias{tidy_words}
\title{Description}
\usage{
tidy_words(collection, data_dir = "data")
}
\arguments{
\item{data_ref_authors}{a tibble produced with the tidy_ref_authors() function}
}
\value{
a tibble with all title words
}
\description{
Description
}
\examples{
data=extract_collection("EVS_UMR5600", nmax=50)
data_ref_authors=tidy_ref_authors(data)
data_words=tidy_words(data_)
}
