% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Cross people or labs to quantify number of co-publications}
\usage{
cross(custom_name, data_dir, type = "people")
}
\arguments{
\item{custom_name}{a publications list custom name}

\item{data_dir}{the path to data directory}

\item{type}{whether to group by "people" or "labs"}

\item{method}{method to shorten names. Defaults to "shortest"}
}
\value{
a tibble which crosses people or labs to tally number of collaborations
}
\description{
Cross people or labs to quantify number of co-publications
}
\examples{
HALtere::cross("BIOEENVIS","data_HALtere", type="people")
}
