% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hal_labels.R
\name{get_hal_labels}
\alias{get_hal_labels}
\title{Retrieve human-readable labels from HAL identifiers}
\usage{
get_hal_labels(
  identifier,
  hal_variable = c("anrProjectReference_s", "collCode_s", "authIdHal_s",
    "structAcronym_s")
)
}
\arguments{
\item{identifier}{identifier to query}

\item{type}{one of:
- "project"
- "author"
- "lab"
- "collection"}
}
\value{
a tibble with identifiers and labels
}
\description{
Retrieve human-readable labels from HAL identifiers
}
