% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_all_data.R
\name{prepare_all_data}
\alias{prepare_all_data}
\title{Get all HALtere data}
\usage{
prepare_all_data(collection, hal_variable = "collection")
}
\arguments{
\item{collection}{a collection name}

\item{hal_variable}{the HAL variable used to retrieve items. Can be collCode_s (for collections), authIdHal_s (for person)}
}
\value{
the characteristics of all datasets generated for the collection. Can be "collection", "person", "lab".
}
\description{
Get all HALtere data
}
\examples{
prepare_all_data("lise-vaudor",hal_variable="person")
}
