% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_authors.R
\name{tidy_groups}
\alias{tidy_groups}
\title{Get groups stats from publications and groups-relative stats}
\usage{
tidy_groups(
  custom_name,
  data_dir = "data",
  method = "shortest",
  type = "people"
)
}
\arguments{
\item{custom_name}{a publications list custom name}

\item{data_dir}{the path to data directory}

\item{method}{method to shorten names. Defaults to "shortest"}

\item{type}{whether to group by "people" or "labs"}
}
\value{
a tibble
}
\description{
Get groups stats from publications and groups-relative stats
}
\examples{
data_labs=tidy_groups("BIOEENVIS",type="labs")
}
